requirements.txt
# Core
anthropic>=0.40.0
pyyaml>=6.0

# PDF generation
reportlab>=4.1
pypdf>=4.0

# Article extraction
trafilatura>=1.12

# Email - Gmail
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

# Email - Outlook
msal>=1.28
requests>=2.31

# HTML to text (for HTML emails)
html2text>=2024.2

# Testing
pytest>=8.0
pytest-asyncio>=0.23
